DRAW-198 refactor: ws dto 패키지 수정
