--- app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/request/dto/MafiaAnswerRequest.kt
@@ -1,5 +1,0 @@
-package com.xorker.draw.websocket.message.request.dto
-
-data class MafiaAnswerRequest(
-    val answer: String,
-)
