--- app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/request/dto/VoteMafiaRequest.kt
@@ -1,7 +1,0 @@
-package com.xorker.draw.websocket.message.request.dto
-
-import com.xorker.draw.user.UserId
-
-data class VoteMafiaRequest(
-    val userId: UserId,
-)
