+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/request/dto/game/MafiaGameVoteMafiaRequest.kt
@@ -1,0 +1,5 @@
+package com.xorker.draw.websocket.message.request.dto.game
+
+data class MafiaGameVoteMafiaRequest(
+    val userId: Long,
+)
