+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/game/MafiaGameAnswerMessage.kt
@@ -1,15 +1,15 @@
-package com.xorker.draw.websocket.message.response.dto
+package com.xorker.draw.websocket.message.response.dto.game
 
 import com.xorker.draw.websocket.ResponseAction
 import com.xorker.draw.websocket.SessionMessage
 
-class MafiaAnswerMessage(
-    override val body: MafiaAnswerBody,
+class MafiaGameAnswerMessage(
+    override val body: MafiaGameAnswerBody,
 ) : SessionMessage {
     override val action = ResponseAction.ANSWER
     override val status = SessionMessage.Status.OK
 }
 
-data class MafiaAnswerBody(
+data class MafiaGameAnswerBody(
     val answer: String,
 )
