+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/game/MafiaGameVoteStatusMessage.kt
@@ -1,16 +1,16 @@
-package com.xorker.draw.websocket.message.response.dto
+package com.xorker.draw.websocket.message.response.dto.game
 
 import com.xorker.draw.user.UserId
 import com.xorker.draw.websocket.ResponseAction
 import com.xorker.draw.websocket.SessionMessage
 
-class MafiaVoteStatusMessage(
-    override val body: MafiaVoteStatusBody,
+class MafiaGameVoteStatusMessage(
+    override val body: MafiaGameVoteStatusBody,
 ) : SessionMessage {
     override val action = ResponseAction.VOTE_STATUS
     override val status = SessionMessage.Status.OK
 }
 
-data class MafiaVoteStatusBody(
+data class MafiaGameVoteStatusBody(
     val players: Map<UserId, List<UserId>>,
 )
