+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/phase/MafiaPhaseInferAnswerMessage.kt
@@ -2,7 +2,6 @@ package com.xorker.draw.websocket.message.response.dto.phase
 
 import com.xorker.draw.websocket.ResponseAction
 import com.xorker.draw.websocket.SessionMessage
-import com.xorker.draw.websocket.message.response.dto.MafiaGameInfoMessage
 import java.time.LocalDateTime
 
 class MafiaPhaseInferAnswerMessage(
