+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/phase/MafiaPhaseVoteMessage.kt
@@ -3,7 +3,6 @@ package com.xorker.draw.websocket.message.response.dto.phase
 import com.xorker.draw.user.UserId
 import com.xorker.draw.websocket.ResponseAction
 import com.xorker.draw.websocket.SessionMessage
-import com.xorker.draw.websocket.message.response.dto.MafiaGameInfoMessage
 import java.time.LocalDateTime
 
 class MafiaPhaseVoteMessage(
