+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/dto/phase/MafiaPhaseWaitMessage.kt
@@ -3,8 +3,7 @@ package com.xorker.draw.websocket.message.response.dto.phase
 import com.xorker.draw.room.RoomId
 import com.xorker.draw.websocket.ResponseAction
 import com.xorker.draw.websocket.SessionMessage
-import com.xorker.draw.websocket.message.response.dto.MafiaGameOptionResponse
-import com.xorker.draw.websocket.message.response.dto.MafiaPlayerResponse
+import com.xorker.draw.websocket.message.response.dto.game.MafiaPlayerResponse
 
 data class MafiaPhaseWaitMessage(
     override val body: MafiaPhaseWaitBody,
