+++ domain/src/main/kotlin/com/xorker/draw/mafia/MafiaGameMessenger.kt
@@ -4,11 +4,8 @@ import com.xorker.draw.room.RoomId
 
 interface MafiaGameMessenger {
     fun broadcastPlayerList(gameInfo: MafiaGameInfo)
-    fun broadcastGameInfo(mafiaGameInfo: MafiaGameInfo)
-    fun broadcastGameReady(mafiaGameInfo: MafiaGameInfo)
-    fun broadcastPlayerTurnList(mafiaGameInfo: MafiaGameInfo)
     fun broadcastDraw(roomId: RoomId, data: Map<String, Any>)
     fun broadcastNextTurn(gameInfo: MafiaGameInfo)
-    fun broadcastVoteStatus(mafiaGameInfo: MafiaGameInfo)
+    fun broadcastVoteStatus(gameInfo: MafiaGameInfo)
     fun broadcastAnswer(gameInfo: MafiaGameInfo, answer: String)
 }
